Moved all styles from Workouts.js to stylesWorkouts.js
