+++ components/stylesWorkouts.js
@@ -1,18 +1,5 @@
 import styled from "styled-components";
-
-import React from "react";
-
 import { StyleSheet } from "react-native";
-
-import {
-  View,
-  Text,
-  Image,
-  TextInput,
-  TouchableOpacity,
-  RecyclerViewBackedScrollViewComponent,
-  EdgeInsetsPropType,
-} from "react-native";
 
 // colors
 export const Colors = {
@@ -25,19 +12,141 @@ export const Colors = {
   red: red,
 };
 
-const { primary, secondary, tertiary, darkLight, brand, green, red } = Colors;
+const { red } = Colors;
 
 export const TitleBox = styled.View``;
 
-const styledRectangle = StyleSheet.create({
-  rectangle: {
-    height: 35,
-    width: 130,
-    backgroundColor: "#515151",
-    position: "absolute",
-    marginTop: 100,
-    marginLeft: 0,
-    //left: "10%",
-    opacity: 0.5,
+const modalDialog = StyleSheet.create({
+  centeredView: {
+    flex: 1,
+    justifyContent: "center",
+    alignItems: "center",
+    marginTop: 22,
+  },
+  modalView: {
+    backgroundColor: "white",
+    borderRadius: 20,
+    width: "85%",
+  },
+  modalTitle: {
+    fontSize: 22,
+    fontWeight: "500",
+  },
+  titleView: {
+    width: "100%",
+    alignItems: "center",
+    marginTop: 20,
+  },
+  workoutNameInput: {
+    fontSize: 15,
+    height: 40,
+    marginTop: 15,
+    paddingLeft: 10,
+    width: "85%",
+    marginLeft: 22,
+    borderWidth: 1.5,
+    borderColor: "rgba(255, 0, 0, 0.5)",
+    borderRadius: 10,
+  },
+  typeStyle: {
+    fontSize: 15,
+    marginRight: 10,
+  },
+  typeViewStyle: {
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "center",
+    marginTop: 15,
+  },
+  selectStyle: {},
+  dropDownPicker: {},
+  buttonContainer: {
+    flexDirection: "row",
+    marginTop: 15,
+    padding: 5,
+    width: "100%",
+    borderBottomRightRadius: 20,
+    borderBottomLeftRadius: 20,
+    justifyContent: "space-evenly",
+    borderTopWidth: 0.2,
+  },
+  cancelButtonStyle: {},
+  submitButtonStyle: {},
+});
+
+const styles = StyleSheet.create({
+  image: {
+    flex: 1,
+    width: "100%",
+    height: "100%",
+    justifyContent: "center",
+  },
+  menuBarStyle: {
+    flexDirection: "column",
+    backgroundColor: "rgba(0, 0, 0, 1)",
+    height: 75,
+    width: "100%",
+    justifyContent: "space-between",
+  },
+  menuIconStyle: {
+    flexDirection: "row",
+    padding: 5,
+    justifyContent: "space-evenly",
+  },
+  workoutNameInput: {
+    color: "rgba(255, 255, 255, 1)",
+    fontSize: 25,
+    height: 40,
+    fontWeight: "bold",
+    marginLeft: 10,
   },
 });
+
+const styledRectangle = StyleSheet.create({
+  bigRectangle: {
+    width: 360,
+    marginTop: 10,
+    marginLeft: 0,
+    borderRadius: 28,
+    padding: 10,
+    backgroundColor: "rgba(81, 81, 81, 0.35)",
+    alignSelf: "center",
+    borderWidth: 2,
+    borderColor: "red",
+  },
+  smallRectangle: {
+    marginTop: 15,
+    marginLeft: 0,
+    borderBottomLeftRadius: 0,
+    borderTopLeftRadius: 0,
+    borderTopRightRadius: 10,
+    borderBottomRightRadius: 10,
+    paddingTop: 5,
+    paddingBottom: 5,
+    paddingLeft: 25,
+    paddingRight: 10,
+    backgroundColor: "rgba(81, 81, 81, 0.5)",
+    width: 130,
+  },
+});
+
+const appFontStyles = StyleSheet.create({
+  titleText: {
+    fontFamily: "System",
+    fontSize: 30,
+    marginTop: 45,
+    color: "white",
+  },
+  workoutInfoStyle: {
+    flexDirection: "column",
+    marginTop: 10,
+    marginLeft: 10,
+  },
+  workoutInfoText: {
+    color: "red",
+    fontSize: 15,
+    fontFamily: "System",
+  },
+});
+
+export { modalDialog, styles, styledRectangle, appFontStyles };
